comment 조회 수정 중

+++ src/main/java/com/example/triviewer/vote/service/VoteService.java
@@ -91,7 +91,7 @@ public class VoteService {
             }
         }
 
-        int commentVoteCount = comment.getCommentLikeCount();
+        long commentVoteCount = comment.getCommentLikeCount();
         commentVoteCount += commentVoteStatus;
         comment.setCommentLikeCount(commentVoteCount);
         commentRepository.save(comment);
